### 2021-12-02 #3964 @kthyng
@emdupre would this be an ok fit for you from the imaging direction?
