### 2021-12-06 #3964 @danielskatz
Given your willingness, I will go ahead and assign you - feel free to reach out to other JOSS editors for suggestions, if needed.   Thanks!!
