### 2021-12-07 #3964 @emdupre
Great, thanks @danielskatz @kthyng !

:wave: Hi @po60nani, and thank you for your submission to JOSS !

If you have any suggestions for potential reviewers, please let me know by listing their names or GitHub handles (without the @, so they don't receive a notification) here. You can suggest reviewers from any relevant project, though we often recommend starting with [this list of people](https://bit.ly/joss-reviewers) as they have already agreed to review for JOSS (please start at the bottom of the list).
