### 2021-12-13 #3964 @caldarolamartin
Hi Elizabeth,

Thanks for considering me for the review, but since I am on leave I am
afraid I can not review the paper at this time. If you still need another
reviewer I suggest Aquiles Carattino, his background is similar to mine.

Thanks for understanding, I hope I can review next time.

Best,
Martin

-------------------------

Martín Caldarola

On Mon, 13 Dec 2021 at 01:28, Elizabeth DuPre ***@***.***>
wrote:

> Hello @caldarolamartin <https://github.com/caldarolamartin> @ziatdinovmax
> <https://github.com/ziatdinovmax> @jonatanalvelid
> <https://github.com/jonatanalvelid> @emmanuelle
> <https://github.com/emmanuelle> 👋
>
> Would you be willing to review *PiSCAT: A Python Package for
> Interferometric Scattering Microscopy* for JOSS (the Journal of Open
> Source Software) <https://joss.theoj.org> ?
>
> The review takes place on GitHub and focuses on the software
> <https://github.com/SandoghdarLab/PiSCAT> and this short paper:
> https://github.com/openjournals/joss-papers/blob/joss.03964/joss.03964/10.21105.joss.03964.pdf
>
> If you're available to review this work, please let me know and I'll add
> you as a reviewer. Once I have sufficient reviewers I'll open a dedicated
> review issue.
>
> Thank you !
>
