### 2021-12-14 #3964 @emdupre
Hello @aquilesC :wave:

Would you be willing to review _PiSCAT: A Python Package for Interferometric Scattering Microscopy_ for [JOSS (the Journal of Open Source Software)](https://joss.theoj.org/)?

The review takes place on GitHub and focuses on [the software](https://github.com/SandoghdarLab/PiSCAT) and this short paper: https://github.com/openjournals/joss-papers/blob/joss.03964/joss.03964/10.21105.joss.03964.pdf

If you're available to review this work, please let me know and I'll add you as a reviewer. Once I have sufficient reviewers I'll open a dedicated review issue.

Thank you !
