### 2021-12-15 #3964 @emdupre
Thank you, @aquilesC !

Yes, that timeline is very reasonable (both since JOSS is still in [reduced service mode](https://blog.joss.theoj.org/2020/05/reopening-joss) and since we're quickly approaching the holiday season !). I'll go ahead and add you as a reviewer now :rocket:
