### 2021-12-21 #4024 @whedon
[REVIEW]: PiSCAT: A Python Package for Interferometric Scattering Microscopy
**Submitting author:** @po60nani (<a href="http://orcid.org/0000-0002-0724-3158">Houman Mirzaalian Dastjerdi</a>)
**Repository:** <a href="https://github.com/SandoghdarLab/PiSCAT" target ="_blank">https://github.com/SandoghdarLab/PiSCAT</a>
**Version:** 0.1.20
