### 2021-12-21 #4024 @emdupre
Hi everyone :wave: Thanks again @ziatdinovmax and @aquilesC for agreeing to review this submission ! The review will take place in this issue.

Whenever possible, please open relevant issues on the linked software repository (and cross-link them with this issue) rather than discussing them here. This helps to make sure that feedback is translated into actionable items to improve the software. If you aren't sure how to get started, please see the [Reviewing for JOSS guide](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) -- and, of course, feel free to reach out with any questions !
