### 2021-12-21 #4024 @emdupre
@po60nani a few initial points on the JOSS paper itself, rather than the software :
- [ ] The [manuscript](https://github.com/openjournals/joss-papers/blob/joss.04024/joss.04024/10.21105.joss.04024.pdf) does not yet have a 'Statement of Need' section. You can find more information in the [submission guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) on what the statement of need section should contain
- [ ] Several references cited in the paper are missing DOIs, with whedon suggesting possible additions in [this comment](https://github.com/openjournals/joss-reviews/issues/4024#issuecomment-999013573). Could you please confirm whether the automatically suggested DOIs are appropriate for your references, and either add or correct them ?
