+++ docs/paper.bib
@@ -7,6 +7,7 @@
   pages = {037401},
   numpages = {4},
   year = {2004},
+  doi = {10.1103/PhysRevLett.93.037401}
 }
 
 @article{taylor2019interferometric,
@@ -17,16 +18,18 @@
   number={8},
   pages={4827--4835},
   year={2019},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi = {10.1021/acs.nanolett.9b01822}
 }
 
-@inbook{RichardBook, 
-title={Interferometric Scattering (iSCAT) Microscopy and Related Techniques}, booktitle={Label-Free Super-Resolution Microscopy}, 
-publisher={Springer International Publishing}, 
-author={Taylor, Richard W. and Sandoghdar, Vahid}, 
-year={2019}, 
-pages={25–65},
-chapter={2}
+@inbook{RichardBook,
+    title={Interferometric Scattering (iSCAT) Microscopy and Related Techniques}, booktitle={Label-Free Super-Resolution Microscopy},
+    publisher={Springer International Publishing},
+    author={Taylor, Richard W. and Sandoghdar, Vahid},
+    year={2019},
+    pages={25–65},
+    chapter={2},
+    doi = {10.1007/978-3-030-21722-8}
 }
 
 @article{Dastjerdi2021,
@@ -37,7 +40,8 @@ chapter={2}
   number={5},
   pages={054002},
   year={2021},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi = {10.1088/1361-6463/ac2f68}
 }
 
 @article{piliarik2014direct,
@@ -48,7 +52,9 @@ chapter={2}
   number={1},
   pages={1--8},
   year={2014},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi = {10.1038/ncomms5495}
+
 }
 
 @article{spindler2018high,
@@ -59,7 +65,8 @@ chapter={2}
   number={8},
   pages={5262--5271},
   year={2018},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi = {10.1021/acs.nanolett.8b02240}
 }
 
 @article{gholami2020iPSF,
