### 2022-01-12 #4024 @aquilesC
I am going through the program, and I'll be incrementally listing what I find (some are possible improvements, some are important things to address), I'll clean them up at the end of the process:

It is GREAT that you provide the software and sample data to get started! Kudos!

- [ ] Couldn't find a simple way of downloading sample data. The ``download_tutorial_data`` failed because of write access (was trying to write to ``/Tutorials`` instead of a relative path. When looked into the code, I found the link to the ownclodu zip file. This could be linked from the docs themselves.
- [ ] A bit of documentation on the files I'll be downloading is also helpful. Right now I got two .raw files without contextual information, which means I can't use the GUI to open them (it asks for information that it does not discover on the file, and I have no quick way of telling what it actually is, like image width and heigh)
