### 2022-01-17 b253246
adding GUIBatchAnalysis
