+++ piscat/GUI/BatchAnalysis/tab_reading_data.py
@@ -1,0 +1,177 @@
+from piscat.GUI.InputOutput import Reading
+from functools import partial
+
+
+from PySide2 import QtCore
+from PySide2 import QtWidgets
+
+import os
+import numpy as np
+
+
+class LoadData_GUI(QtWidgets.QWidget):
+
+    update_output_internal = QtCore.Signal(object)
+    update_output_external = QtCore.Signal(object)
+
+    update_tab_index = QtCore.Signal(int)
+
+    def __init__(self):
+        super(LoadData_GUI, self).__init__()
+        self.original_video = None
+        self.filename = None
+        self.folder = None
+        self.data_in = None
+
+        self.update_output_internal.connect(self.updata_setting)
+
+        self.info_video_setting = None
+        self.empty_value_box_flag = False
+
+        self.load = QtWidgets.QPushButton("load video")
+        self.load.setAutoDefault(False)
+        self.load.clicked.connect(self.read_data)
+        self.load.setEnabled(True)
+        self.load.setFixedWidth(100)
+
+        self.Next = QtWidgets.QPushButton("Next")
+        self.Next.setAutoDefault(False)
+        self.Next.clicked.connect(self.update_tab)
+        self.Next.setEnabled(True)
+        self.Next.setFixedWidth(100)
+
+        self.LineEdit_path = QtWidgets.QTextEdit(self)
+        self.LineEdit_path_label = QtWidgets.QLabel("Path:")
+        self.LineEdit_path.setFixedHeight(20)
+        self.LineEdit_path.setFixedWidth(200)
+
+        self.LineEdit_type = QtWidgets.QLineEdit()
+        self.LineEdit_type.setPlaceholderText(".bin")
+        self.LineEdit_type_label = QtWidgets.QLabel("Data type:")
+        self.LineEdit_type.setFixedHeight(20)
+        self.LineEdit_type.setFixedWidth(200)
+
+        self.name_mkdir = QtWidgets.QLineEdit()
+        self.name_mkdir.setPlaceholderText("folder_name")
+        self.name_mkdir_label = QtWidgets.QLabel("Save folder name:")
+        self.name_mkdir.setFixedHeight(20)
+        self.name_mkdir.setFixedWidth(200)
+
+        self.LineEdit_img_width = QtWidgets.QTextEdit(self)
+        self.LineEdit_img_width_label = QtWidgets.QLabel("img_width:")
+        self.LineEdit_img_width.setFixedHeight(20)
+        self.LineEdit_img_width.setFixedWidth(200)
+
+        self.LineEdit_img_height = QtWidgets.QTextEdit(self)
+        self.LineEdit_img_height_label = QtWidgets.QLabel("img_height:")
+        self.LineEdit_img_height.setFixedHeight(20)
+        self.LineEdit_img_height.setFixedWidth(200)
+
+        self.LineEdit_img_type = QtWidgets.QTextEdit(self)
+        self.LineEdit_img_type_label = QtWidgets.QLabel("image_type:")
+        self.LineEdit_img_type.setFixedHeight(20)
+        self.LineEdit_img_type.setFixedWidth(200)
+
+        self.LineEdit_frame = QtWidgets.QTextEdit(self)
+        self.LineEdit_frame_label = QtWidgets.QLabel("frame:")
+        self.LineEdit_frame.setFixedHeight(20)
+        self.LineEdit_frame.setFixedWidth(200)
+
+        self.grid = QtWidgets.QGridLayout()
+        self.grid.addWidget(self.createFirstExclusiveGroup(), 0, 0)
+
+
+        self.setLayout(self.grid)
+
+    def createFirstExclusiveGroup(self):
+        groupBox = QtWidgets.QGroupBox("File browser:")
+
+        self.grid_file_browser = QtWidgets.QGridLayout()
[…]
