+++ piscat/GUI/BatchAnalysis/tab_plot_histogram.py
@@ -1,0 +1,164 @@
+from piscat.Analysis.plot_protein_histogram import PlotProteinHistogram
+
+from PySide2 import QtGui, QtCore, QtWidgets
+
+
+class Histogram_GUI(QtWidgets.QWidget):
+    update_tracking = QtCore.Signal(object)
+    output_setting_Tab_tracking = QtCore.Signal(object)
+
+    def __init__(self, parent=None):
+        super(Histogram_GUI, self).__init__(parent)
+
+        self.input_data = None
+        self.lower_limit = None
+        self.upper_limit = None
+        self.max_n_components = None
+        self.minPeak_width = None
+        self.n_bins = None
+
+        self.empty_value_box_flag = False
+
+        self.step_limit = 1e-6
+        self.setting_tracking = {}
+
+        self.resize(300, 300)
+        self.setWindowTitle('Histogram')
+
+        self.plot_hist = QtWidgets.QPushButton("plot_histogram")
+        self.plot_hist.clicked.connect(self.do_hist)
+        self.plot_hist.setFixedWidth(150)
+
+        self.le_lower_contrast_trim = QtWidgets.QLineEdit()
+        self.le_lower_contrast_trim.setPlaceholderText("lower_limitation")
+        self.le_lower_contrast_trim_label = QtWidgets.QLabel("lower limitation (Contrast):")
+        self.le_lower_contrast_trim.setFixedWidth(150)
+
+        self.le_upper_contrast_trim = QtWidgets.QLineEdit()
+        self.le_upper_contrast_trim.setPlaceholderText("upper_limitation")
+        self.le_upper_contrast_trim_label = QtWidgets.QLabel("upper limitation (Contrast):")
+        self.le_upper_contrast_trim.setFixedWidth(150)
+
+        self.le_MaxGMM_mode = QtWidgets.QLineEdit()
+        self.le_MaxGMM_mode.setPlaceholderText("max_n_components for GMM")
+        self.le_MaxGMM_mode_label = QtWidgets.QLabel("max n_components:")
+        self.le_MaxGMM_mode.setFixedWidth(150)
+
+        self.le_hist_bin = QtWidgets.QLineEdit()
+        self.le_hist_bin.setPlaceholderText("#bins")
+        self.le_hist_bin_label = QtWidgets.QLabel("Histogram bins:")
+        self.le_hist_bin.setFixedWidth(150)
+
+        self.min_peak_width = QtWidgets.QLineEdit()
+        self.min_peak_width.setPlaceholderText("Min Peak Width")
+        self.min_peak_width_label = QtWidgets.QLabel("Min Peak Width:")
+        self.min_peak_width.setFixedWidth(150)
+
+        self.axis_limit = QtWidgets.QLineEdit()
+        self.axis_limit.setPlaceholderText("scale * upper_limitation")
+        self.le_axis_limit = QtWidgets.QLabel("Scale x-axis:")
+        self.axis_limit.setFixedWidth(150)
+
+        self.GMM_visualization_mode_group = QtWidgets.QButtonGroup()
+        self.radio_mode_1 = QtWidgets.QRadioButton("Separate mixture")
+        self.radio_mode_2 = QtWidgets.QRadioButton("Superposition mixture")
+        self.GMM_visualization_mode_group.addButton(self.radio_mode_1)
+        self.GMM_visualization_mode_group.addButton(self.radio_mode_2)
+
+        self.grid = QtWidgets.QGridLayout()
+        self.grid.addWidget(self.createFirstExclusiveGroup(), 0, 0)
+        self.grid.addWidget(self.createSecondExclusiveGroup(), 1, 0)
+        self.grid.addWidget(self.plot_hist, 2, 0)
+
[…]
