### 2022-01-17 bee395f
adding GUIBatchAnalysis
