+++ docs/requirements.txt
@@ -28,3 +28,4 @@ h5py
 ipykernel
 m2r2
 recommonmark
+sphinx-markdown-tables
