### 2022-01-17 909af8b
adding GUIBatchAnalysis
