+++ CONTRIBUTING.md
@@ -7,7 +7,7 @@ Of course, it is also always appreciated to bring own ideas and problems to the 
 
 Please submit all contributions to the official [Github repository](https://github.com/SandoghdarLab/PiSCAT/) in the form of a Merge Request. Please do not submit git diffs or files containing the changes.
 
-`PiSCAT` is an open-source python package under the license of [GPLv3](https://github.com/SandoghdarLab/PiSCAT/blob/main/LICENSE). Thus we consider the act of contributing to the code by submitting a Merge Request as the "Sign off" or agreement to the GPLv3 license.
+`PiSCAT` is an open-source python package under the license of [GNUv3](https://github.com/SandoghdarLab/PiSCAT/blob/main/LICENSE). Thus we consider the act of contributing to the code by submitting a Merge Request as the "Sign off" or agreement to the GNUv3 license.
 
 You can contribute in many different ways:
 
@@ -35,12 +35,12 @@ Ready to contribute? Here is how to set up `PiSCAT` for local development.
 1. Fork the `PiSCAT` repo on GitHub.
 2. Clone your fork locally:
 ```bash
-    $ git clone https://github.com/SandoghdarLab/PiSCAT.git
+    $ git clone https://github.com/USERNAME/PiSCAT.git
 ```
 3. Install your local copy into a virtualenv. It is also recommended to use anaconda or miniconda to manage the python environments.
 ```bash
-    $ mkvirtualenv PiSCAT
-    $ cd PiSCAT/
+    $ conda create -n PiSCAT python=3.9
+    $ conda activate PiSCAT
     $ pip install -e .
 ```
 4. Create a branch for local development:
@@ -49,8 +49,7 @@ Ready to contribute? Here is how to set up `PiSCAT` for local development.
 ```
    Now you can make your changes locally.
 
-5. When you're done making changes, check that your changes pass flake8 and the
-   tests
+5. When you're done making changes, check that your changes pass the tests
 ```bash
     $ python setup.py test" .
 ```
@@ -69,7 +68,7 @@ Ready to contribute? Here is how to set up `PiSCAT` for local development.
 
 Before you submit a Merge Request, check that it meets these guidelines:
 
-1. All functionality that is implemented through this Merge Request should be covered by unit tests. These are implemented in `PiSCAT_tests`. It would be necessary to add your unit test if the new implementation has some features that are not covered by our unit tests.
+1. All functionality that is implemented through this Merge Request should be covered by unit tests. These are implemented in `PiSCAT\tests`. It would be necessary to add your unit test if the new implementation has some features that are not covered by our unit tests.
 2. If the Merge Request adds functionality, the docs should be updated. Put your new functionality into a function with a docstring.
 3. If you have a maintainer status for `PiSCAT`, you can merge Merge Requests to the master branch. However, every Merge Request needs to be reviewed by another developer. Thus it is not allowed to merge a Merge Request, which is submitted by oneself.
 
