### 2022-02-02 #4024 @ziatdinovmax
My apologies for the delay in reviewing the package. I was able to install it successfully and ran examples following the instructions provided by the authors. Their manuscript is well-written: it explains the need for this package and clearly outlines its structure. At the end of the manuscript, the authors mention future plans such as GPU support and incorporation of DNN. I would expect there are people who may be willing to contribute, but the contributing guidelines seem to be absent. Can the authors add them to the README file?

One other minor issue: I recommend specifying version requirements for the packages listed in setup.py (open-cv, networkx, etc) as some of the listed packages have undergone quite a few breaking changes between different releases.
