### 2022-02-11 a060bbb
adding requirements.txt
