+++ requirements.txt
@@ -1,0 +1,23 @@
+numpy==1.20.3
+scipy==1.7.3
+matplotlib==3.5.0
+opencv-python==4.5.4.60
+pandas==1.3.4
+PyQt5==5.15.6
+PySide2==5.15.2
+pyqtgraph==0.12.3
+PyQtWebEngine==5.15.5
+scikit-learn==1.0.1
+scikit-image==0.18.3
+tqdm==4.62.3
+trackpy==0.5.0
+joblib==1.1.0
+psutil==5.8.0
+ipywidgets==7.6.5
+networkx==2.6.3
+numba==0.54.1
+requests==2.26.0
+wget==3.2
+h5py==3.6.0
+imageio-ffmpeg==0.4.5
+astropy==5.0
