### 2022-02-11 #4024 @po60nani
**I was able to install it successfully and ran examples following the instructions provided by the authors. Their manuscript is well-written: it explains the need for this package and clearly outlines its structure.**

- _We appreciate the reviewer's acknowledgment of our paper's and package's utility and effectiveness._

**At the end of the manuscript, the authors mention future plans such as GPU support and incorporation of DNN. I would expect there are people who may be willing to contribute, but the contributing guidelines seem to be absent. Can the authors add them to the README file?**

- _We've already added this section to the community guidelines, and we've also added a link to contributing in the [Readme](https://github.com/SandoghdarLab/PiSCAT/blob/main/README.md#contributing), based on the second reviewer's comments._

**One other minor issue: I recommend specifying version requirements for the packages listed in setup.py (open-cv, networkx, etc) as some of the listed packages have undergone quite a few breaking changes between different releases.**

- _In our repository, we added a file called requirements.txt that specifies version requirements. We also mentioned this in the contributing guidelines, along with the Python version we tested._
