### 2022-03-08 #4024 @aquilesC
I have finished reviewing the manuscript. The only concern I have is regarding the **references**, and I would ask @emdupre for guidance on the matter.

The authors claim:

> We would have to mention all dependencies and explain them in the text if we wanted to add a Trackpy citation, which would be problematic because we intended to keep the number of pages under four.

At the same time, in the manuscript they say:

> PiSCAT provides several algorithms for localization and tracking of nanoparticles

They do implement several localisation algorithms, but only one for tracking and it is based on Trackpy.

Generally speaking, the PiSCAT package has [23 dependencies](https://github.com/SandoghdarLab/PiSCAT/blob/main/requirements.txt), some with clear academic relevance (Astropy, Trackpy, PyQtGraph) and some larger ones with scientific relevance (numpy, SciPy, scikit, matplotlib). I wonder where the threshold to require a citation for software should be put. After all, there are 7 self-references, but none to software.

Once this is clarified *editorially* (also for my own reference for future reviews), I'll conclude the review.
