### 2022-03-11 971b67c
Pyside2 has been upgraded to version 6 with better display features.
