+++ piscat/GUI/Projects/tab_trajectory.py
@@ -2,7 +2,7 @@ from piscat.Trajectory import particle_linking, temporal_filtering
 from piscat.Visualization import plot, plot_histogram
 from piscat.Analysis.plot_protein_histogram import PlotProteinHistogram
 
-from PySide2 import QtGui, QtCore, QtWidgets
+from PySide6 import QtGui, QtCore, QtWidgets
 
 
 class Tracking_GUI(QtWidgets.QWidget):
