+++ piscat/GUI/InputOutput/import_RAW.py
@@ -1,7 +1,7 @@
 import matplotlib.pyplot as plt
 from piscat.InputOutput import reading_videos
-from PySide2 import QtCore
-from PySide2 import QtWidgets
+from PySide6 import QtCore
+from PySide6 import QtWidgets
 
 
 class RawImage(QtWidgets.QMainWindow):
