+++ piscat/GUI/Memory/video_in_memory.py
@@ -1,11 +1,11 @@
 import psutil
 import gc
 
-from PySide2 import QtGui, QtCore, QtWidgets
-from PySide2.QtWidgets import *
-from PySide2 import QtCore, QtGui
-from PySide2.QtGui import *
-from PySide2.QtCore import *
+from PySide6 import QtGui, QtCore, QtWidgets
+from PySide6.QtWidgets import *
+from PySide6 import QtCore, QtGui
+from PySide6.QtGui import *
+from PySide6.QtCore import *
 from functools import partial
 
 
