### 2022-03-11 4b05466
Pyside2 has been upgraded to version 6 with better display features.
