+++ setup.py
@@ -44,19 +44,19 @@ setup(
                         "matplotlib",
                         "opencv-python",
                         "pandas",
-                        "PyQt5",
+                        "PyQt6",
                         "Pyside6",
                         "pyqtgraph",
                         "PyQtWebEngine",
                         "scikit_learn",
                         "scikit-image",
                         "tqdm",
-                        "trackpy",
                         "joblib",
                         "psutil",
                         "ipywidgets",
                         "networkx",
                         "numba",
+                        "trackpy",
                         "requests",
                         "wget",
                         "h5py",
