+++ requirements.txt
@@ -3,14 +3,13 @@ scipy==1.7.3
 matplotlib==3.5.0
 opencv-python==4.5.4.60
 pandas==1.3.4
-PyQt5==5.15.6
-PySide2==5.15.2
-pyqtgraph==0.12.3
+PyQt6==6.2.3
+PySide6==6.2.3
+pyqtgraph==0.12.4
 PyQtWebEngine==5.15.5
 scikit-learn==1.0.1
 scikit-image==0.18.3
 tqdm==4.62.3
-trackpy==0.5.0
 joblib==1.1.0
 psutil==5.8.0
 ipywidgets==7.6.5
@@ -21,3 +20,5 @@ wget==3.2
 h5py==3.6.0
 imageio-ffmpeg==0.4.5
 astropy==5.0
+trackpy==0.5.0
+
