+++ docs/paper.bib
@@ -83,6 +83,21 @@ year = {2020},
 doi = {10.1364/OE.401374},
 }
 
+@software{Trackpy,
+  author       = {Allan, Daniel B. and
+                  Caswell, Thomas and
+                  Keim, Nathan C. and
+                  van der Wel, Casper M. and
+                  Verweij, Ruben W.},
+  title        = {soft-matter/trackpy: Trackpy v0.5.0},
+  month        = apr,
+  year         = 2021,
+  publisher    = {Zenodo},
+  version      = {v0.5.0},
+  doi          = {10.5281/zenodo.4682814},
+  url          = {https://doi.org/10.5281/zenodo.4682814}
+}
+
 
 
 
