### 2022-03-21 #4024 @emdupre
Thank you for your feedback, @aquilesC, and your recent updates @po60nani ! I'm glad that compiling with the new `editorial-bot` system was smooth :slightly_smiling_face: The previous Whedon system has been deprecated, so note that all future compilation, etc, commands can be issued directly to editorial-bot for both this and any other JOSS submissions.

For clarity, I would just like to confirm that JOSS does not have a strict page limit. Although [we generally recommend](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) that papers are at most approximately 1000 words, this is not a hard limit and can be adjusted to accommodate information that the authors or reviewers believe is necessary to understand the presented research software and its contribution to the field.

With that in mind, I agree with @aquilesC that citing the appropriate software is well-worth inclusion in the paper. I appreciate that the authors have added a reference to `TrackPy` in the "Concept and Structure of PiSCAT" section to point to supported localisation algorithms as a step in this direction. Considering citing other software dependencies:

> Generally speaking, the PiSCAT package has [23 dependencies](https://github.com/SandoghdarLab/PiSCAT/blob/main/requirements.txt), some with clear academic relevance (Astropy, Trackpy, PyQtGraph) and some larger ones with scientific relevance (numpy, SciPy, scikit, matplotlib). I wonder where the threshold to require a citation for software should be put. After all, there are 7 self-references, but none to software.

There is not a strict rule that all dependencies for a given package must be cited. Instead, this is generally an editorial judgment, with more significant dependencies for a project receiving acknowledgement in-text. For example, in the [example paper provided in the JOSS documentation](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography), there is a reference to one other research software project, from which the example project significantly benefits:

> `Gala` also relies heavily on and interfaces well with the implementations of physical units and astronomical coordinate systems in the `Astropy` package [@\astropy] (`astropy.units` and `astropy.coordinates`).

This acknowledgement features in the Statement of Need section, which is generally used to overview the current landscape of the field and how the present package contributes to it. While the authors provide a thorough and effective scope of PiSCAT in the current paper's Statement of Need, it would likely help orient readers if they could directly note there any other research software whose functionality significantly supports or overlaps with that of PiSCAT. It is likely that TrackPy merits mention here, given the usage of its tracking algorithm.

I hope this helps to clarify your questions around references, @aquilesC, and provides relevant context for making appropriate edits, @po60nani ! I have several other, minor edits on the paper and reference formatting which I will include in my next comment.
