### 2022-03-21 #4024 @emdupre
There are a few minor corrections in references that would significantly improve the formatting of the paper.

- **Updating inclusion of multiple references:** For example,  this reference list on L13:
  _"...microsecond temporal resolution over arbitrarily long measurement times [@\lindfors2004PRL], [@\taylor2019interferometric], [@\RichardBook]. "_
  should be updated as:
  _"...microsecond temporal resolution over arbitrarily long measurement times [@\lindfors2004PRL; @\taylor2019interferometric; @\RichardBook]."_
- **Conflicting bracket-parentheses placement on L45:** Here,
  _"...such as lipid vesicles across the surface of giant unilamellar vesicles (GUVs) [@\spindler2018high](see \autoref{fig:piscat}a-(IV))"_
  the missing space between the Spindler and figure reference is yielding inappropriate formatting. I would recommend updating as follows:
  _"...such as lipid vesicles across the surface of giant unilamellar vesicles (GUVs; @\spindler2018high; see also \autoref{fig:piscat}a-(IV))"_
