+++ docs/paper.md
@@ -98,14 +98,14 @@ height is swept axially to form an axial stack made of 2D iPSF images. This stac
 plane and plotted in (II). The amplitude and differential phase of the same plane is shown in (III) and (IV) respectively. 
 \label{fig:piscat}](Fig1.png){ width=100% }
 
-PiSCAT provides several algorithms for localization and tracking [@Trackpy] of nanoparticles (see \autoref{fig:piscat}a-(VI)). The 
+PiSCAT provides several algorithms for localization and tracking of nanoparticles (see \autoref{fig:piscat}a-(VI)). The 
 current version offers several image processing techniques to mainly accomplish the following: 1) Background correction 
 that improves the distinction of the signal of interest from the background, 2) Localizing particles in (X,Y,Z axis and time). 
 To achieve this, we need the combination of different computer vision and machine learning algorithms with physical 
 models such as the iSCAT point-spread function (iPSF) model (see \autoref{fig:piscat}b-(I-IV)) [@gholami2020iPSF]. 
 \autoref{fig:piscat_structure} depicts the structure of the PiSCAT package. 
-PiSCAT uses a set of dependencies that assist us in achieving various goals. The PiSCAT 
-repository contains this list.
+PiSCAT uses a set of dependencies (e.g. PyQtGraph [@pyqtgraph] , scikit-learn/image [@scikit-learn], Trackpy [@Trackpy] 
+and Astropy [@astropy:2018]) that assist us in achieving various goals. The PiSCAT repository contains this list.
 
 ![Diagram of the PiSCAT structure: There are nine main modules in the PiSCAT package (blue boxes). Green and yellow 
 represent classes and python scripts (functions), respectively. 
