+++ docs/paper.bib
@@ -171,7 +171,6 @@ archivePrefix = {arXiv},
  year={2011}
 }
 
- @misc{pyqtgraph, title={Scientific Graphics and GUI library for Python}, url={https://www.pyqtgraph.org/}, journal={PyQtGraph}}
 
 
 
