### 2022-03-22 #4024 @po60nani
@emdupre, Thank you for the comprehensive explanation. We changed the manuscript based on your explanation and comments regarding reference corrections.

- We added the following sentence to clarify the PiSCAT dependencies:

_`PiSCAT uses a set of dependencies (e.g. PyQtGraph, scikit-learn/image [@scikit-learn], Trackpy [@Trackpy], and Astropy [@astropy:2018]) that assist us in achieving various goals. The PiSCAT repository contains this list.`_

- There are a few minor corrections in references that would significantly improve the formatting of the paper.
 _`Done`_
