### 2022-03-28 #4024 @emdupre
Thank you, @aquilesC ! Confirming that I've spoken to @ziatdinovmax via email, who has also confirmed that he's happy to recommend PiSCAT for publication in JOSS. Thank you both for your thoughtful reviews !

I'll perform a few last checks now, but at this point, @po60nani, could you please:

- [ ] Check the archival deposit (e.g., in Zenodo) has the correct metadata (e.g., the JOSS paper and software archive have the same title and author listing)
- [ ] List the DOI of the archived version here.

I can then move forward with accepting the submission :rocket:
