+++ piscat/GUI/Projects/iPSF/iPSF_model.py
@@ -21,121 +21,145 @@ class GUI_iPSF(QtWidgets.QWidget):
         self.empty_value_ModelBox_flag = True
 
         self.wavelength = QtWidgets.QLineEdit()
-        self.wavelength.setPlaceholderText("Wavelength in meters")
+        self.wavelength.setToolTip('Wavelength of the light source in meters')
+        self.wavelength.setPlaceholderText("Wavelength of the light source in meters")
         self.wavelength_label = QtWidgets.QLabel("wavelength:")
         self.wavelength.setFixedWidth(150)
 
         self.NA = QtWidgets.QLineEdit()
-        self.NA.setPlaceholderText("objective lens")
+        self.NA.setToolTip('Numerical Aperture (NA) of the objective lens')
+        self.NA.setPlaceholderText("Numerical Aperture (NA) of the objective lens")
         self.NA_label = QtWidgets.QLabel("Numerical Aperture (NA):")
         self.NA.setFixedWidth(150)
 
         self.ti0 = QtWidgets.QLineEdit()
-        self.ti0.setPlaceholderText("nominal value in meters")
-        self.ti0_label = QtWidgets.QLabel("Thickness of the immersion oil:")
+        self.ti0.setToolTip('Thickness of the immersion oil, nominal value in meters')
+        self.ti0.setPlaceholderText("Thickness of the immersion oil, nominal value in meters")
+        self.ti0_label = QtWidgets.QLabel("Thickness:")
         self.ti0.setFixedWidth(150)
 
         self.ni0 = QtWidgets.QLineEdit()
-        self.ni0.setPlaceholderText("nominal value")
-        self.ni0_label = QtWidgets.QLabel("Refractive index of the immersion oil:")
+        self.ni0.setToolTip('Refractive index of the immersion oil, nominal value')
+        self.ni0.setPlaceholderText("Refractive index of the immersion oil, nominal value")
+        self.ni0_label = QtWidgets.QLabel("Refractive index:")
         self.ni0.setFixedWidth(150)
 
         self.ni = QtWidgets.QLineEdit()
-        self.ni.setPlaceholderText("experimental value")
-        self.ni_label = QtWidgets.QLabel("Refractive index of the immersion oil:")
+        self.ni.setToolTip('Refractive index of the immersion oil, experimental value')
+        self.ni.setPlaceholderText("Refractive index of the immersion oil, experimental value")
+        self.ni_label = QtWidgets.QLabel("Refractive index:")
         self.ni.setFixedWidth(150)
 
         self.tg0 = QtWidgets.QLineEdit()
-        self.tg0.setPlaceholderText("nominal value in meters")
-        self.tg0_label = QtWidgets.QLabel("Thickness of the coverglass:")
+        self.tg0.setToolTip('Thickness of the coverglass, nominal value in meters')
+        self.tg0.setPlaceholderText("Thickness of the coverglass, nominal value in meters")
+        self.tg0_label = QtWidgets.QLabel("Thickness:")
         self.tg0.setFixedWidth(150)
 
         self.tg = QtWidgets.QLineEdit()
-        self.tg.setPlaceholderText("experimental value in meters")
-        self.tg_label = QtWidgets.QLabel("Thickness of the coverglass:")
+        self.tg.setToolTip('Thickness of the coverglass, experimental value in meters')
+        self.tg.setPlaceholderText("Thickness of the coverglass, experimental value in meters")
+        self.tg_label = QtWidgets.QLabel("Thickness:")
         self.tg.setFixedWidth(150)
 
         self.ng0 = QtWidgets.QLineEdit()
-        self.ng0.setPlaceholderText("nominal value")
[…]
