+++ docs/appendix.rst
@@ -7,14 +7,14 @@ Summary of the PiSCAT pipeline for efficient nanoparticle detection
 The implementation flowchart for detecting single unlabeled proteins from iSCAT videos is shown in the following Fig [1].
 
 .. image:: ./Fig/flowchart.png
-   :width: 600 px
-   :alt: PiSCAT pipeline [1]
-   :align: center
+  :width: 400
+  :alt:  PiSCAT pipeline [1]
+  :align: center
 
 
 PiSCAT publications
 -------------------
-[1] Dastjerdi, Houman Mirzaalian, et al. "Optimized analysis for sensitive detection and analysis of single proteins via interferometric scattering microscopy." bioRxiv (2021).
+[1] Dastjerdi, Houman Mirzaalian, et al. "Optimized analysis for sensitive detection and analysis of single proteins via interferometric scattering microscopy." Journal of Physics D: Applied Physics 55.5 (2021): 054002. (`Journal <https://iopscience.iop.org/article/10.1088/1361-6463/ac2f68>`_)
 
 
 
