### 2022-03-28 #4024 @po60nani
Thank you @emdupre.  We updated the repository:

-  **Make a tagged release of your software, and list the version tag of the archived version here.**
  _PiSCAT version 0.1.23 ([link](https://github.com/SandoghdarLab/PiSCAT/releases/tag/0.1.23))_

-  **Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)**
 _Done!_

-  **Check the archival deposit (e.g., in Zenodo) has the correct metadata (e.g., the JOSS paper and software archive have the same title and author listing)**
 _Done! Here you can find the [Zendo](https://zenodo.org/record/6389882#.YkGRRtIzZaQ)_

-  **List the DOI of the archived version here.**

1. [DOI: 10.5281/zenodo.6389603](https://doi.org/10.5281/zenodo.6389603)
2. [DOI: 10.5281/zenodo.6389882](https://doi.org/10.5281/zenodo.6389882)
