### 2022-03-28 #4024 @emdupre
Thank you, @po60nani ! So, to confirm, [10.5281/zenodo.6389882](https://doi.org/10.5281/zenodo.6389882) is the archive for the latest release. I will only associate one DOI with the JOSS publication, so please let me know if the other, earlier archive has non-overlapping information.

On the archive itself, a few formatting requests:
- [ ] Could you please update the title of the archive to match the software paper ? Currently, the archive title appears cut-off (i.e., is missing the term "microscopy" compared to the software paper).
- [ ] Similarly, could you confirm the author listing in the software archive ? The first two authors are differently listed between the two objects (e.g., the first author is `Houman` in the software archive and `Houman Mirzaalian Dastjerdi` in the software paper).

These edits can be made directly on the archived version, without minting a new DOI ! A guide is available [here](https://help.zenodo.org/#technical).
