### 2022-03-30 #4024 @kyleniemeyer
Hi @po60nani, I'm doing some final checks before publishing. Just a few things:
- in the statement of need section, could you change "e.g" to "e.g.," ?
- The journal name in the Dastjerdi et al. reference looks off, could you correct that?
