+++ docs/paper.bib
@@ -35,7 +35,7 @@
 @article{Dastjerdi2021,
   title={Optimized analysis for sensitive detection and analysis of single proteins via interferometric scattering microscopy},
   author={Dastjerdi, Houman Mirzaalian and Dahmardeh, Mahyar and Gemeinhardt, Andr{\'e} and Mahmoodabadi, Reza Gholami and K{\"o}stler, Harald and Sandoghdar, Vahid},
-  journal={J. Phys. D J PHYS D APPL PHYS},
+  journal={J. Phys. D},
   volume={55},
   number={5},
   pages={054002},
