### 2022-03-31 #4024 @kyleniemeyer
@po60nani thanks—but unfortunately there is another issue that I missed, the Astropy paper is missing the journal name (we see this a lot because .bib files often have journal name macros that our system doesn't interpret). Could you fix that too?
