+++ docs/paper.bib
@@ -143,7 +143,7 @@ doi = {10.1364/OE.401374},
          {Weaver}, B.~A. and {Whitmore}, J.~B. and {Woillez}, J. and
          {Zabalza}, V. and {Astropy Contributors}},
         title = "{The Astropy Project: Building an Open-science Project and Status of the v2.0 Core Package}",
-      journal = {\aj},
+      journal = {Astronomy \& Astrophysics},
      keywords = {methods: data analysis, methods: miscellaneous, methods: statistical, reference systems, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2018,
         month = sep,
